+++ lib/pscom4velo/pscom_extoll.c
@@ -106,7 +106,7 @@ int pscom_extoll_velo2_do_read(pscom_poll_reader_t *reader)
 static
 void pscom_extoll_velo2_do_write(pscom_con_t *con)
 {
-	unsigned int len;
+	size_t len;
 	struct iovec iov[2];
 	pscom_req_t *req;
 
@@ -268,7 +268,7 @@ int pscom_extoll_rma2_do_read(pscom_poll_reader_t *reader)
 static
 void pscom_extoll_rma_do_write(pscom_con_t *con)
 {
-	unsigned int len;
+	size_t len;
 	struct iovec iov[2];
 	pscom_req_t *req;
 
@@ -278,7 +278,7 @@ void pscom_extoll_rma_do_write(pscom_con_t *con)
 		psex_con_info_t *ci = con->arch.extoll.ci;
 		len = iov[0].iov_len + iov[1].iov_len;
 
-		int rlen = psex_sendv(ci, iov, len);
+		ssize_t rlen = psex_sendv(ci, iov, len);
 
 		if (rlen >= 0) {
 			pscom_write_done(con, req, rlen);
