+++ lib/pscom4velo/psextoll.h
@@ -102,7 +102,7 @@ void psex_recvdone(psex_con_info_t *con_info);
  * -EAGAIN if ci is busy or
  * -EPIPE in case of a broken connection.
  */
-int psex_sendv(psex_con_info_t *con_info, struct iovec *iov, int size);
+ssize_t psex_sendv(psex_con_info_t *con_info, struct iovec *iov, size_t size);
 
 
 /* Suggest a value for psex_pending_tokens. Result depends on psex_recvq_size. */
@@ -119,7 +119,7 @@ unsigned psex_pending_tokens_suggestion(void);
  * -EAGAIN if ci is busy or
  * -EPIPE in case of a broken connection.
  */
-int psex_velo2_sendv(psex_con_info_t *con_info, struct iovec *iov, int size);
+int psex_velo2_sendv(psex_con_info_t *con_info, struct iovec *iov, size_t size);
 
 
 /* recv up to msg_len bytes into msg. return bytes read or
